Put Java's implicit conversion where Calcite puts it
JavaCast was three of Calcite's things and one of ours under a name Calcite has no counterpart for.

To(expression, type) is EnumUtils.convert(operand, toType) -- the same job, Java's implicit conversion written out because an expression tree will not do it -- so it is ClrEnumUtils.Convert now, beside the rest of that class. Its private Box and Unbox over expressions are Expressions.box and Expressions.unbox, and the widening is what EnumUtils.convert does inside itself. PrimitiveOf was Primitive.ofBox before I wrote it this afternoon.

Unwrap and Box over values are the one part with no counterpart: Java boxes a value in the language, so Calcite never needs a method for it, and linq4j's Primitive offers only array helpers. Those go to JavaValues, which is where this port already converts a value crossing between the runtimes.

11 core, 103 adapter, 275 data, 321 Linq.
